REST_API/requirements.txt
django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
django-filter==22.1
drf-spectacular==0.23.1
drf-yasg==1.21.3
gunicorn==20.1.0
whitenoise==6.2.0